Update function_app.py
really too many changes to go into detail of what changed since extensive testing and monitoring was done in my private development repo to the point where I'm comfortable pushing everything to this repo which is being treated as a pseudo-production repo with my workflow :
summary:
implemented OpenTelemetry custom monitoring & logging since the default Azure Application Insights logging wasn't sufficient enough for me to debug with & OpenTelemetry will be the backbone of Azure logging in the future. 

Changed the execution order of function calls to further safeguard against unhandled exceptions and to use a Twitter response Tweet confirmation as proof of work so that entries into CosmosDB can be made with 100% certainty that a Tweet was made with the content being inserted.

Tweaked the AI generated image being tweeted to reduce errors in spelling, grammar, formatting, and other issues that a human normally wouldn't make. 
	This implements the use of an additional Azure AI feature.